+++ packages/data-explorer/src/repository/dataRepository.ts
@@ -6,6 +6,7 @@ import { MetaData, Attribute } from '@/types/MetaData'
 import client from '@/lib/client'
 import { encodeRsqlValue } from '@molgenis/rsql'
 import { Pagination } from '@molgenis-ui/components-library'
+import { Sort } from '@/types/Sort'
 
 // maps api response to object with as key the name of the column and as value the label of the value or a list of labels for mrefs
 const levelOneRowMapper = async (rowData: DataApiResponseItem, metaData: MetaData) => {
@@ -94,7 +95,7 @@ const addFilterIfSet = (request: string, rsqlFilter?: string): string => {
   return rsqlFilter ? `${request}&q=${encodeRsqlValue(rsqlFilter)}` : request
 }
 
-const pageQuery = (pagination :Pagination|undefined) => {
+const buildPageQuery = (pagination :Pagination|undefined) => {
   // (!) Adapt page to 0-based data-api page property.
   return pagination ? Object
     .entries({ page: pagination.page - 1, size: pagination.size })
@@ -103,39 +104,55 @@ const pageQuery = (pagination :Pagination|undefined) => {
     : ''
 }
 
+const buildSortQuery = (sort?: Sort):string => {
+  return sort && sort.sortColumnName !== null ? `&sort=${sort.isSortOrderReversed ? '-' : ''}${sort.sortColumnName}` : ''
+}
+
 const getTableDataDeepReference = async (
   tableId: string,
   metaData: MetaData,
-  coloms: string[],
+  columns: string[],
   rsqlQuery?: string,
-  pagination?: Pagination
+  pagination?: Pagination,
+  sort?: Sort
 ) => {
-  if (!coloms.includes(metaData.idAttribute.name)) {
-    coloms.push(metaData.idAttribute.name)
+  if (!columns.includes(metaData.idAttribute.name)) {
+    columns.push(metaData.idAttribute.name)
   }
 
-  if (metaData.labelAttribute !== undefined && !coloms.includes(metaData.labelAttribute.name)) {
-    coloms.push(metaData.labelAttribute.name)
+  if (metaData.labelAttribute !== undefined && !columns.includes(metaData.labelAttribute.name)) {
+    columns.push(metaData.labelAttribute.name)
   }
 
-  const expandReferencesQuery = buildExpandedAttributesQuery(metaData, coloms)
+  const expandReferencesQuery = buildExpandedAttributesQuery(metaData, columns)
 
-  const request = addFilterIfSet(`/api/data/${tableId}?${pageQuery(pagination)}${expandReferencesQuery}`, rsqlQuery)
+  const sortQuery = buildSortQuery(sort)
+  const pageQuery = buildPageQuery(pagination)
+  const request = addFilterIfSet(`/api/data/${tableId}?${pageQuery}${expandReferencesQuery}${sortQuery}`, rsqlQuery)
   const response = (await client.get<DataApiResponse>(request)).data
 
   const items = response.items.map((item: DataApiResponseItem) => levelNRowMapper(item))
   return { items, page: response.page }
 }
 
-const getTableDataWithLabel = async (tableId: string, metaData: MetaData, columns: string[], rsqlQuery?: string, pagination?: Pagination) => {
+const getTableDataWithLabel = async (
+  tableId: string,
+  metaData: MetaData,
+  columns: string[],
+  rsqlQuery?: string,
+  pagination?: Pagination,
+  sort?: Sort
+) => {
   const columnSet = new Set([...columns])
   columnSet.add(metaData.idAttribute.name)
   if (metaData.labelAttribute !== undefined) {
     columnSet.add(metaData.labelAttribute.name)
   }
 
+  const sortQuery = buildSortQuery(sort)
   const expandReferencesQuery = buildExpandedAttributesQuery(metaData, [...columnSet])
-  const request = addFilterIfSet(`/api/data/${tableId}?${pageQuery(pagination)}${expandReferencesQuery}`, rsqlQuery)
+  const pageQuery = buildPageQuery(pagination)
+  const request = addFilterIfSet(`/api/data/${tableId}?${pageQuery}${expandReferencesQuery}${sortQuery}`, rsqlQuery)
   const response = (await client.get<DataApiResponse>(request)).data
   // @ts-ignore
   const items = await Promise.all(response.items.map(async (item: DataApiResponseItem) => {
@@ -146,7 +163,7 @@ const getTableDataWithLabel = async (tableId: string, metaData: MetaData, column
 }
 
 // called on row expand
-const getRowDataWithReferenceLabels = async (tableId: string, rowId: string, metaData: MetaData, pagination?: Pagination) => {
+const getRowDataWithReferenceLabels = async (tableId: string, rowId: string, metaData: MetaData, pagination?: Pagination, sort?: Sort) => {
   const attributes: string[] = getAttributesfromMeta(metaData)
   // Todo: remove work around, needed as compounds are not passed by getAttributesfromMeta.
   // Adding id and label makes sure we get these fields.
@@ -155,8 +172,10 @@ const getRowDataWithReferenceLabels = async (tableId: string, rowId: string, met
   if (metaData.labelAttribute !== undefined) {
     columnSet.add(metaData.labelAttribute.name)
   }
+  const sortQuery = buildSortQuery(sort)
+  const pageQuery = buildPageQuery(pagination)
   const expandReferencesQuery = buildExpandedAttributesQuery(metaData, [...columnSet])
-  const response = await client.get<DataApiResponse>(`/api/data/${tableId}/${rowId}?${pageQuery(pagination)}${expandReferencesQuery}`)
+  const response = await client.get<DataApiResponse>(`/api/data/${tableId}/${rowId}?${pageQuery}${expandReferencesQuery}${sortQuery}`)
   return levelOneRowMapper(response.data, metaData)
 }